added submit new foods frond end view, route and controller, moved foodModel.js into app for gulp finding and made changes to server.js
