+++ src/app/views/pages/home.html
@@ -22,6 +22,8 @@
       <li><p>Description</p></li>
     </ul>
   </div>
+  <!-- commented markup below is for later use -->
+  
 <!--   <div class="col" ng-repeat="awesomeThing in main.awesomeThings | orderBy:'rank'">
     <div class="thumbnail">
       <img class="pull-right" ng-src="assets/images/{{ awesomeThing.logo }}" alt="{{ awesomeThing.title }}">
